WIP: Snake moves on keypress
At the moment the snake is only moving left, I suspect this is happening
because they KEY_LEFT & KEY_RIGHT codes are not the same from getch().

What I need is that the program continues to run without the need of
user input. Meaning that the snake should continue moving in whatever
direction it was set until another key is pressed.