Remove Threads - Moving Snake & User Input
This commit removes the previous implementation, which used Threads,
although it was not working as expected as the user input was not being
executed straight away.

The solution was to use nodelay(), just before using getch().
I thought that having it in the screenSetup() function was more than
enough, seems I was wrong.

Ive taken the oportunity to actually move the Snake using WASD keyboard
keys and instead of increment/decrement snakeDirection variable I've now
changed it to an enum instead, it's easier to read and easier to play
the game when you know which direction you want to take.
With this in mind I also took in consideration the user input being in
the oposite direction that the snake is going and prevented such thing
to happen.

BUG:

The snake tail is still in place from the start of the game, which means
that the snake is growing without eating any fruit. I'm still to
understand how to erase it as the snake moves around.